requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-anthropic>=0.1.0
anthropic>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
